Renamed ContentDescriptor -> DescriptorGenerator
This new name is more descriptive of what the interface and
implementations aim to do. The previous name can also be confused with
a data type, which is what the DescriptorElement interface is intende
to encapsulate.
