+++ python/smqtk/algorithms/descriptor_generator/colordescriptor/INSTALL.md
