+++ python/smqtk/algorithms/descriptor_generator/colordescriptor/__init__.py
