+++ python/smqtk/algorithms/descriptor_generator/colordescriptor/encode_FLANN.py
