+++ python/smqtk/algorithms/descriptor_generator/colordescriptor/utils.py
